third_party/WebKit/LayoutTests/fast/css/usecounter-position3value-basic-shape-ellipse.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let ThreeValuedPositionBasicShape = 2002; // From UseCounter.h

    let isCounted = () => internals.isUseCounted(document, ThreeValuedPositionBasicShape);
    var div = document.createElement('div');

    // usecounter-position3value-basic-shape-circle.html
    // tests that other position-using properties do not
    // increment the counter.

    div.style = 'shape-outside: ellipse(at left);';
    div.style = 'shape-outside: ellipse(at left 10%);';
    div.style = 'shape-outside: ellipse(at left 10% top 20%);';
    assert_false(isCounted(),
                 '1,2,4 values should not be counted');

    div.style = 'shape-outside: ellipse(at left 10% top);';
    assert_true(isCounted(),
                'basic-shape should be counted');
}, 'Three valued position syntax is use counted for ellipse');
</script>
